sample.anova.test {MSTAT}	R Documentation

Sample size for One-way ANOVA

Description

Gives the required minimum sample size for t test

Usage

sample.anova.test(k, d, sd, w = NULL, sig.level = 0.05, power = 0.8,
  nonpar = F, inflate = c(0.15, 0.2))

Arguments

k
No. of groups (No.of multiple comparisons will be calculated automatically)

d
The clinically significant difference

sd
The estimate of standard deviation

w
The effect size

sig.level
The significance level defaults to 0.05

power
Power of the test defaults to 0.8(80%)

nonpar
Should be set as nonpar=T if nonparametric inflation is needed.

inflate
a vector or single value specifying the % of inflation needed for nonparametric inflation default output is for 15% and 20%

Author(s)

Nithin Nayak

[Package MSTAT version 0.1.0 Index]