sample.prop.test {MSTAT}	R Documentation

Sample size for comparison of proportion

Description

Sample size for comparison of proportion

Usage

sample.prop.test(p1, p2, d = NULL, sig.level = 0.05, power = 0.8,
  oneside = F)

Arguments

d
The expected difference in proportion. If not entered d will be calculated as (p2-p1)

oneside
Enter T if the test is onesided.

p1&p2
estimate of proportion of first and second group respectively

sig.level&power
Significance level and power of the test respectively.

Value

Returns the minimum sample size requierd per group

Author(s)

Nithin Nayak

[Package MSTAT version 0.1.0 Index]