sample.t.test {MSTAT}	R Documentation

Sample size for t test

Description

Gives the required minimum sample size for t test

Usage

sample.t.test(d = NULL, sd = NULL, w = NULL, sig.level = 0.05,
  power = 0.8, oneside = F, nonpar = F, inflate = c(0.2, 0.15),
  paired = F)

Arguments

d
The clinically significant difference

sd
The estimate of standard deviation

w
The effect size

sig.level
The significance level defaults to 0.05

power
Power of the test defaults to 0.8(80%)

oneside
Should set oneside=T if the test is onesided

nonpar
Should be set as nonpar=T if nonparametric inflation is needed.

inflate
a vector or single value specifying the % of inflation needed for nonparametric inflation default output is for 15% and 20%

paired
Should set paired=T if sample size is for paired t test

Author(s)

Nithin Nayak

[Package MSTAT version 0.1.0 Index]